% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Source.R
\docType{data}
\name{DF_Seg_Chile}
\alias{DF_Seg_Chile}
\title{Segregation data in southern Chile}
\format{
A data.frame with 191495 observations and 11 variables:
\describe{
\item{year}{Academic year. From 2016 to 2018.}
\item{school}{School ID (RBD, Rol de Base de Datos).}
\item{commune}{Official commune code of the school location.}
\item{csep}{Socio-economic criterion. Students belong to either the non-vulnerable (1), the preferent (2), or the
priority (3) group acording to the Act 20.248 of Preferencial Scholar Subsidy (SEP).}
\item{ethnicity}{Ethnicity criterion. Students belong to mapuche ethnicity (1) or not (0).}
\item{rural}{Rurality criterion. The school is located in a urban zone (0) or not (1).}
\item{region}{Official region code of the school location. Schools can belong either Bio-Bío region (8), La Araucanía
region (9) or Los Ríos region (14).}
\item{sch_type}{Dependency code. School dependancy as municipal (1), private subsidized (2) or private paid (3).}
\item{gender}{Binary student gender. Students can either be female (1) or male (2).}
\item{grade}{Student grade. Students can either belong to the 4th (4) or 8th (8) grade of basic school.}
\item{nobs}{Number of students in a cell or combination of variables.}
}
}
\source{
<https://github.com/RafaelFuentealbaC/mutualinf>

<http://datosabiertos.mineduc.cl/>

<http://www.agenciaeducacion.cl/>
}
\usage{
DF_Seg_Chile
}
\description{
The data set included in this package was build using two data sets. The first one is the student
enrollment reported by the Ministry of Education (MINEDUC) <http://datosabiertos.mineduc.cl/> for students of primary
education (first eight years of formal education) who attended establishments officially recognized by the State.
The second one is the Quality and Context of Education Questionnaire for Parents and Guardians, and the Student
Questionnaire, both applied by the Education Quality Agency <http://www.agenciaeducacion.cl/> to all students in
grades 4 and 8 of primary education. Both sources are limited to the period 2016-2018. Contains information related
to students and educational system characteristics in southern Chile (Bio-Bio, La Araucanía and Los Ríos regions).
}
\keyword{datasets}
