% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Source.R
\docType{data}
\name{DT_Seg_Ar}
\alias{DT_Seg_Ar}
\title{Segregation data of the Araucania region}
\format{
A data.table with 20107 observations and 10 variables:
\describe{
\item{year}{school year.}
\item{school}{establishment database role.}
\item{comuna}{official code of commune in which the establishment is located.}
\item{csep}{socio-economic criterion.}
\item{etnia}{belongs to indigenous people.}
\item{rural}{rural RBD.}
\item{region}{region code where the establishment is located.}
\item{sch_type}{dependence code of the establishment.}
\item{gender}{sex of student.}
\item{fw}{number of students.}
}
}
\source{
<https://github.com/RafaelFuentealbaC/mutualinf>
}
\usage{
DT_Seg_Ar
}
\description{
Contain information from estudents of the Araucania Region in Chile between year 2016 and 2018.
The data was be obtain from ... ?
}
\keyword{datasets}
