% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual.R
\name{mutual}
\alias{mutual}
\title{Calculates the mutual information index}
\usage{
mutual(
  data,
  group,
  unit,
  within = NULL,
  by = NULL,
  components = FALSE,
  contribution.from = NULL,
  cores = NULL
)
}
\arguments{
\item{data}{A data.table.}

\item{group}{A categorical variable name or vector of categorical variables names contained in \code{data}, or also,
a categorical variable number or vector of categorical variables numbers contained in \code{data}. Defines the first
dimension over which segregation is computed.}

\item{unit}{A categorical variable name or vector of categorical variables names contained in \code{data}, or also,
a categorical variable number or vector of categorical variables numbers contained in \code{data}. Defines the second
dimension over which segregation is computed.}

\item{within}{A categorical variable name or vector of categorical variables names contained in \code{data}, or also,
a categorical variable number or vector of categorical variables numbers contained in \code{data}. Defines the
dimensions within which segregation is computed. By default is NULL.}

\item{by}{A categorical variable name or vector of categorical variables names contained in \code{data}, or also, a
categorical variable number or vector of categorical variables numbers contained in \code{data}. Defines the
dimensions by which calculations are separated. By default is NULL.}

\item{components}{A boolean value. If is TRUE and the \code{within} option is not null and the \code{by} option
is null then returns a \code{list} where the first element is a \code{data.table} that contains a summary of the index
total value and decompositions while the second element is a \code{data.table} with more detail information of the
decomposition of the within term. Besides if the \code{within} and \code{by} options are not null then returns a
structure of \code{list of lists} type where each first element is a \code{data.table} that contains the summary of
the index total value and decompositions while each second element is a \code{data.table} with more detail information
of the decomposition of the within term that displayed in each first element. The detailed information includes the
proportions and local segregation. By default is FALSE.}

\item{contribution.from}{A variable of character type that can be 'group_vars' or 'unit_vars', or also, a categorical
variable name or vector of categorical variables names contained in the \code{group} parameter or \code{unit}
parameter, or also, a categorical variable number or vector of categorical variables numbers contained in the
\code{group} parameter or \code{unit} parameter. Defines the contributing dimension into the total segregation. By
default is NULL.}

\item{cores}{A positive integer. Defines the amount of CPU cores that is use to parallelization task into the
index compute. If is NULL then the compute is carried out sequentially in only one core. This option is available
to Mac, Linux, Unix, and BSD systems but is not available to Windows sytems. By default is NULL.}
}
\value{
A data.table if the \code{components} option is false; a list if the \code{components} option is true and
the \code{within} option is not null and the \code{by} option is null; or a list of lists if the \code{components}
option is true and the \code{within} and \code{by} options are not null.
}
\description{
Function that delivery the index total value and the decompositions into the between term and the within
term, either for all data, a particular variable, or multiple variables. Besides allows know the local segregation
and calculate the exclusive contribution that generate the group variables or the unit variables into the total
segregation. The decompositions of the within term and contributions can be show in the general form or in detail form.
}
\examples{
\dontrun{
# Get the total segregation.
mutual(data = DT_Seg_Ar, group = c("csep", "etnia"), unit = "school")

# Using the 'by' option to calculate the segregation by year or by another variable(s).
mutual(data = DT_Seg_Ar, group = c("csep", "etnia"), unit = "school", by = "year")

# Use the 'within' option to decompose the index total value into the between term and the within term.
mutual(data = DT_Seg_Ar, group = c("csep", "etnia"), unit = "school", within = "etnia")

# Use the 'components' option to get detail information. The result shows the proportions and the local segregation index
# on the 'W_Decomposition' element. The weighted average between 'p' and 'within' is equal to the within term.
mutual(data = DT_Seg_Ar, group = c("csep", "etnia"), unit = "school", within = "etnia", component = TRUE)

# Use the 'contribution.from' option to get exclusive contributions of the segregation sources into the total segregation.
# Contribution from of all variables of 'group' elements:
mutual(data = DT_Seg_Ar, group = c("csep", "etnia"), unit = "school", by = "year", contribution.from = "group_vars")

# Contribution only from 'etnia' variable:
mutual(data = DT_Seg_Ar, group = c("csep", "etnia"), unit = "school", by = "year", contribution.from = "etnia")

# Use the 'cores' option to increase the CPU cores into the index compute.
mutual(data = DT_Seg_Ar, group = c("csep", "etnia"), unit = "school", by = "year", cores = 2)
}
}
\references{
Frankel, D. & Volij, O. (2011). Measuring school segregation. Journal of Economic Theory. 146(1):1-38. DOI:
10.1016/j.jet.2010.10.008.

Mora, R. & Guinea-Martin, D. (2021). Computing decomposable multigroup indexes of segregation. UC3M Working
papers. Economics 31803, Universidad Carlos III de Madrid. Departamento de Economía.

Mora, R. & Ruiz-Castillo, J. (2003). Additively decomposable segregation indexes. The case of gender segregation by
occupations and human capital levels in spain. Journal of Economic Inequality. 1(2):147-179.
DOI: 10.1023/A:1026198429377.

Mora, R. & Ruiz-Castillo, J. (2011). Entropy-based segregation indices. Sociological Methodology. 41(1):159-194.
DOI: 10.1111/j.1467-9531.2011.01237.x.

Theil, H. & Finizza, Anthony J. (1971). A note on the measurement of racial integration of schools by means of
informational concepts. The Journal of Mathematical Sociology. 1(2):187-193. DOI: 10.1080/0022250X.1971.9989795.
}
